解决无法播放的问题

+++ library/src/main/java/com/gibbs/gplayer/GPlayer.java
@@ -159,7 +159,7 @@ public class GPlayer implements IGPlayer, OnPositionChangedListener {
 
     @Override
     public int getDuration() {
-        return (int) mMediaSource.getDuration();
+        return mMediaSource.getDuration();
     }
 
     @Override
